Move write_arrays and write_dicts to core_io._base_write
