+++ src/geff/graph_libs/_networkx.py
@@ -4,6 +4,7 @@ from typing import TYPE_CHECKING, Any, Literal
 
 import networkx as nx
 
+from geff.core_io import write_dicts
 from geff.core_io._base_read import read_to_memory
 from geff.io_utils import (
     calculate_roi_from_nodes,
@@ -11,7 +12,6 @@ from geff.io_utils import (
     get_graph_existing_metadata,
 )
 from geff.metadata_schema import GeffMetadata, axes_from_lists
-from geff.write_dicts import write_dicts
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
