+++ src/geff/graph_libs/_rustworkx.py
@@ -14,6 +14,7 @@ except ImportError as e:
     ) from e
 
 
+from geff.core_io import write_dicts
 from geff.core_io._base_read import read_to_memory
 from geff.io_utils import (
     calculate_roi_from_nodes,
@@ -21,7 +22,6 @@ from geff.io_utils import (
     get_graph_existing_metadata,
 )
 from geff.metadata_schema import GeffMetadata, axes_from_lists
-from geff.write_dicts import write_dicts
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
