+++ src/geff/graph_libs/_spatial_graph.py
@@ -23,10 +23,10 @@ if TYPE_CHECKING:
     from geff.typing import PropDictNpArray
 
 import geff
+from geff.core_io import write_arrays
 from geff.core_io._base_read import read_to_memory
 from geff.metadata_schema import GeffMetadata, axes_from_lists
 from geff.utils import remove_tilde
-from geff.write_arrays import write_arrays
 
 
 def write_sg(
