+++ tests/test_agnostic/test_dtype_validation.py
@@ -3,9 +3,9 @@ from typing import Any
 import numpy as np
 import pytest
 
+from geff.core_io import write_arrays
 from geff.metadata_schema import GeffMetadata
 from geff.valid_values import validate_data_type
-from geff.write_arrays import write_arrays
 
 
 # -----------------------------------------------------------------------------
